+++ lib/coinone/order/cancel_order_response.rb
@@ -1,0 +1,27 @@
+
+module Coinone
+
+  class Order
+
+    class CancelOrderResponse
+
+      attr_reader :result
+
+      def initialize(options={})
+
+        @result = options[:result] || nil
+
+        update_response(options)
+
+
+      end
+
+      def update_response(params={})
+
+        @result = params[:result] if params.has_key? :result
+
+      end
+
+    end
+  end
+end