+++ test/test_coinone_public.rb
@@ -1,26 +1,43 @@
-require 'helper'
-class TestCoinonePublic < Minitest::Test
-
-  def setup
-
-    @connection = Coinone::Connection.new(access_token: ENV['COINONE_ACCESS_TOKEN'], secret_key: ENV['COINONE_SECRET_KEY'])
-
-  end
+class TestCoinonePublic < TestCoinone
 
   def test_get_currency
-    assert_equal Coinone::Public.get_currency().result, "success"
+
+    response = get_data_body("public/currency")
+    data = JSON.parse(response, symbolize_names: true)
+    currency = Coinone::Public::Currency.new(data)
+
+    assert_equal currency.result, "success"
+
   end
 
   def test_get_orderbook
-    assert_equal Coinone::Public.get_orderbook().result, "success"
+
+    response = get_data_body("public/order_book")
+    data = JSON.parse(response, symbolize_names: true)
+    order_book = Coinone::Public::Currency.new(data)
+
+    assert_equal order_book.result, "success"
+
   end
 
   def test_get_complete_orders
-    assert_equal Coinone::Public.get_complete_orders().result, "success"
+
+    response = get_data_body("public/recent_complete_orders")
+    data = JSON.parse(response, symbolize_names: true)
+    recent_complete_orders = Coinone::Public::Currency.new(data)
+
+    assert_equal recent_complete_orders.result, "success"
+
   end
 
   def test_get_ticker
-    assert_equal Coinone::Public.get_ticker().result, "success"
+
+    response = get_data_body("public/ticker")
+    data = JSON.parse(response, symbolize_names: true)
+    ticker = Coinone::Public::Currency.new(data)
+
+    assert_equal ticker.result, "success"
+
   end
 
 end