+++ CHANGE_LOG.md
@@ -1,5 +1,37 @@
 ## Release History
 
+* **v0.9.0** - 2019-08-30
+  - Update `rest-client` dependency version (https://github.com/rest-client/rest-client/issues/713)
+
+* **v0.8.0.a** - 2017-05-31
+  - Add XRP Code to `PUBLIC / OrderBook`, `PUBLIC / Recent Complete Orders`, `PUBLIC / Ticker` API
+
+* **v0.7.0** - 2017-04-03
+  - Change Code to Ruby Convension( present? to !OBJ.nil?)
+  - Make Test Code
+
+
+* **v0.6.0** - 2017-03-30
+  - Porting `TransactionV2 / 2-Factor Authentication`
+  - Porting `TransactionV2 / Coin TransactionsHistory`
+  - Porting `TransactionV2 / KRW Transactions History`
+  - Porting `TransactionV2 / Send Coin`
+
+* **v0.5.0** - 2017-03-29
+  - Porting `OrderV2 / Cancel All Order`
+  - Porting `OrderV2 / Cancel Order`
+  - Porting `OrderV2 / Limit Buy`
+  - Porting `OrderV2 / Limit Sell`
+  - Porting `OrderV2 / Market Buy`
+  - Porting `OrderV2 / Market Sell`
+  - Porting `OrderV2 / My Complete Orders`
+  - Porting `OrderV2 / My Limit Orders`
+
+
+* **v0.4.0** - 2017-03-15
+  - Porting `AccountV2 / Daily Balance`
+  - Porting `AccountV2 / Deposit Address`
+
 * **v0.3.0** - 2017-03-10
   - Porting `AccountV2 / Balance`
   - Add Response Struct to README.md file